Accept fractional balance trades

+++ sysproduction/interactive_order_stack.py
@@ -199,7 +199,7 @@ def create_balance_trade(data):
     instrument_code, contract_date = get_valid_instrument_code_and_contractid_from_user(
         data
     )
-    fill_qty = get_and_convert("Quantity ", type_expected=int, allow_default=False)
+    fill_qty = get_and_convert("Quantity ", type_expected=float, allow_default=False)
     filled_price = get_and_convert(
         "Filled price", type_expected=float, allow_default=False
     )
